백준: Atrium (250506)

+++ study/src/juyoungoh/beakjoon/_20353.java
@@ -9,7 +9,7 @@ public class _20353 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
-        int a = Integer.parseInt(br.readLine());
+        long a = Long.parseLong(br.readLine());
 
         double answer = 4 * Math.sqrt(a);
         System.out.printf("%.8f\n", answer);
